CHROMIUM: iwl7000: add wide firmware command support for notifications
Add support for extended command id in notification system.
Extended command id header contains group id in addition to command id.

iwl7000-tree: 63d02ba72fa5d407ef4b39e76ceac9a88909b5b1

+++ drivers/net/wireless-3.8/iwl7000/iwlwifi/xvt/fw.c
@@ -168,7 +168,7 @@ static int iwl_xvt_load_ucode_wait_alive(struct iwl_xvt *xvt,
 	const struct fw_img *fw;
 	int ret;
 	enum iwl_ucode_type old_type = xvt->cur_ucode;
-	static const u8 alive_cmd[] = { XVT_ALIVE };
+	static const u16 alive_cmd[] = { XVT_ALIVE };
 
 	xvt->cur_ucode = ucode_type;
 	fw = iwl_get_ucode_image(xvt, ucode_type);
